pyproject.toml
[project]
name = "auth-manager-svc"
version = "1.0.0"
description = "Auth Manager Service - microservice for Keycloak token management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "httpx>=0.28.0",
    "cryptography>=44.0.0",
    "structlog>=24.4.0",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "python-keycloak>=5.8.1",
    "scalar-fastapi>=1.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"

[dependency-groups]
dev = ["ty>=0.0.1a24"]


[tool.ty.environment]
root = ["./app"]

[tool.ty.src]
exclude = ["tests"]
